app/requirements.txt
# determining file mimetype
python-magic
# reading and manipulating PDF files
pypdf2
# interacting with AWS S3 storage
boto3
# reading yaml config file
PyYAML
# parsing docx component files
lxml